# swagger.yaml
openapi: 3.0.3
info:
  title: Movie Booking System API
  description: |
    Multi-city, multi-theatre movie ticket booking platform built on Java 21 + Spring Boot 3.3.5.

    ## Authentication
    Most endpoints require a JWT Bearer token. Obtain one via `POST /api/auth/login` or `POST /api/auth/register`.
    Pass it in the `Authorization` header as `Bearer <token>`.

    ## Roles
    | Role | Access |
    |---|---|
    | **CUSTOMER** | Browse, hold seats, book, cancel, view history |
    | **ADMIN** | Full catalogue management (cities, theatres, screens, movies, shows, pricing, discounts, refund policies) |

    ## Error Shape
    All error responses follow `{ "status": <int>, "message": "<string>", "timestamp": "<ISO-8601>" }`.
  version: 1.0.0
  contact:
    email: manya@adobe.com
servers:
  - url: http://localhost:8080
    description: Local development server
...
security: []   # default: public; overridden per-operation
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Enums ──────────────────────────────────────────────────────────────────
  schemas:
...
    CreateCityRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Mumbai
...
    MovieResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: Interstellar
        description:
          type: string
          example: A team of explorers travel through a wormhole in space.
        duration:
          type: integer
          description: Duration in minutes
          example: 169
        language:
          type: string
          example: English
        genre:
          type: string
          example: Sci-Fi
...
    PagedMovieResponse:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/MovieResponse'
    PagedShowResponse:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/ShowResponse'
...
  responses:
...
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 404
            message: Resource not found
            timestamp: "2025-07-15T17:30:00"
...
  parameters:
    PageParam:
      name: page
      in: query
      description: Zero-based page index (default 0)
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      description: Page size (default 20)
      schema:
        type: integer
        default: 20
        minimum: 1

# ═══════════════════════════════════════════════════════════════════════════════
# Paths
# ═══════════════════════════════════════════════════════════════════════════════
paths:
...
  /api/movies:
    get:
      tags: [Movies]
      summary: List all movies (paginated)
      operationId: getAllMovies
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated list of movies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedMovieResponse'
  /api/movies/{id}:
    get:
      tags: [Movies]
      summary: Get a movie by ID
      operationId: getMovieById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Movie detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Shows (public) ────────────────────────────────────────────────────────
  /api/cities/{cityId}/shows:
    get:
      tags: [Shows]
      summary: List all shows in a city (paginated)
      operationId: getShowsByCity
      parameters:
        - name: cityId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated list of shows in the city
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedShowResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/shows/{id}:
    get:
      tags: [Shows]
      summary: Get show detail by ID
      operationId: getShowById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Show detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/admin/cities:
    post:
      tags: [Admin — Cities]
      summary: Create a city
      operationId: createCity
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCityRequest'
      responses:
        '200':
          description: City created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityResponse'
        '400':
          description: Validation error or city name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags: [Admin — Cities]
      summary: List all cities
      operationId: getAllCities
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of all cities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CityResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/cities/{id}:
    put:
      tags: [Admin — Cities]
      summary: Update a city
      operationId: updateCity
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCityRequest'
      responses:
        '200':
          description: City updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin — Cities]
      summary: Delete a city (cascades to theatres → screens → seats)
      operationId: deleteCity
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: City deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Admin — Theatres ──────────────────────────────────────────────────────
  /api/admin/theatres:
    post:
      tags: [Admin — Theatres]
      summary: Create a theatre
      operationId: createTheatre
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTheatreRequest'
      responses:
        '200':
          description: Theatre created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TheatreResponse'
        '400':
          description: Validation error or city not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags: [Admin — Theatres]
      summary: List all theatres
      operationId: getAllTheatres
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of all theatres
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TheatreResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/theatres/{id}:
    put:
      tags: [Admin — Theatres]
      summary: Update a theatre
      operationId: updateTheatre
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTheatreRequest'
      responses:
        '200':
          description: Theatre updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TheatreResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin — Theatres]
      summary: Delete a theatre (cascades to screens → seats)
      operationId: deleteTheatre
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Theatre deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Admin — Screens ───────────────────────────────────────────────────────
  /api/admin/screens:
    post:
      tags: [Admin — Screens]
      summary: Create a screen
      operationId: createScreen
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScreenRequest'
      responses:
        '200':
          description: Screen created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScreenResponse'
        '400':
          description: Validation error or theatre not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags: [Admin — Screens]
      summary: List all screens
      operationId: getAllScreens
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of all screens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScreenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/screens/{id}:
    put:
      tags: [Admin — Screens]
      summary: Update a screen
      operationId: updateScreen
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScreenRequest'
      responses:
        '200':
          description: Screen updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScreenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin — Screens]
      summary: Delete a screen (cascades to seats)
      operationId: deleteScreen
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Screen deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/screens/{id}/seats:
    post:
      tags: [Admin — Screens]
      summary: Bulk-create seats for a screen
      description: |
        Creates multiple physical seat records for an existing screen.
        Each seat specifies its row, seat number, and type (REGULAR or PREMIUM).
        ShowSeat availability rows are created automatically when a Show is scheduled on this screen.
      operationId: createSeats
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Screen ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkCreateSeatsRequest'
      responses:
        '200':
          description: Seats created; returns the resulting ShowSeat availability records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShowSeatResponse'
        '400':
          description: Validation error or screen not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Admin — Movies ────────────────────────────────────────────────────────
  /api/admin/movies:
    post:
      tags: [Admin — Movies]
      summary: Create a movie
      operationId: createMovie
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMovieRequest'
      responses:
        '200':
          description: Movie created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/movies/{id}:
    put:
      tags: [Admin — Movies]
      summary: Update a movie
      operationId: updateMovie
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMovieRequest'
      responses:
        '200':
          description: Movie updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin — Movies]
      summary: Delete a movie
      operationId: deleteMovie
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Movie deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Admin — Shows ─────────────────────────────────────────────────────────
  /api/admin/shows:
    post:
      tags: [Admin — Shows]
      summary: Schedule a show
      description: |
        Creates a show for a given movie on a given screen.
        Guard: the screen must not have another show overlapping the requested time window
        (`startTime`–`endTime`). Overlapping shows on the same screen are rejected with HTTP 400.
      operationId: createShow
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShowRequest'
      responses:
        '200':
          description: Show scheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowResponse'
        '400':
          description: Validation error or overlapping show on the same screen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Movie or screen not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/shows/{id}:
    put:
      tags: [Admin — Shows]
      summary: Update a show
      operationId: updateShow
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShowRequest'
      responses:
        '200':
          description: Show updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin — Shows]
      summary: Delete a show
      operationId: deleteShow
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Show deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Admin — Pricing Tiers ─────────────────────────────────────────────────
  /api/admin/pricing-tiers:
    post:
      tags: [Admin — Pricing Tiers]
      summary: Create a pricing tier
      description: |
        A pricing tier maps a SeatType × DayType combination to a base price and multiplier.
        Effective price = `basePrice × multiplier` (scale 2, HALF_UP).
        The four standard tiers are REGULAR_WEEKDAY, REGULAR_WEEKEND, PREMIUM_WEEKDAY, PREMIUM_WEEKEND.
      operationId: createPricingTier
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePricingTierRequest'
      responses:
        '200':
          description: Pricing tier created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingTierResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags: [Admin — Pricing Tiers]
      summary: List all pricing tiers
      operationId: getAllPricingTiers
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of all pricing tiers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PricingTierResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/pricing-tiers/{id}:
    put:
      tags: [Admin — Pricing Tiers]
      summary: Update a pricing tier
      operationId: updatePricingTier
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePricingTierRequest'
      responses:
        '200':
          description: Pricing tier updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingTierResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin — Pricing Tiers]
      summary: Delete a pricing tier
      operationId: deletePricingTier
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Pricing tier deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Admin — Discount Codes ────────────────────────────────────────────────
  /api/admin/discount-codes:
    post:
      tags: [Admin — Discount Codes]
      summary: Create a discount code
      description: |
        - **FLAT** discount: subtracts a fixed amount; result floored at zero.
        - **PERCENTAGE** discount: `totalAmount × discountValue / 100`, scale 2, HALF_UP.
        - At most one discount code may be applied per booking.
      operationId: createDiscountCode
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDiscountCodeRequest'
      responses:
        '200':
          description: Discount code created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscountCodeResponse'
        '400':
          description: Validation error or code already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags: [Admin — Discount Codes]
      summary: List all discount codes
      operationId: getAllDiscountCodes
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of all discount codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiscountCodeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/discount-codes/{id}:
    put:
      tags: [Admin — Discount Codes]
      summary: Update a discount code
      operationId: updateDiscountCode
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDiscountCodeRequest'
      responses:
        '200':
          description: Discount code updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscountCodeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin — Discount Codes]
      summary: Delete a discount code
      operationId: deleteDiscountCode
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Discount code deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Admin — Refund Policies ───────────────────────────────────────────────
...